use atomic seq for subscriber
both the replay goroutine and the live-tail goroutine read and write from a subscriber's `seq` property concurrently, and before this change that can lead to data races, reproducible locally by `go run -race`, and observable typically as out-of-order events emitted to clients around cutover from replay to live-tailing.

this change uses an atomic int64 to protect the subscriber `seq`. however, this only addresses the physical memory access bug. both goroutines can still end up doing an unsynchronized read-modify-write of the sequence, and events can still be emitted out of order.

the *consequences* of this race are improved though: removed some undefined behaviour.

additionally, an atomic `swap` is used when writing the new value, which allows at least detecting that the race has occurred. right now we just log a warning for that.

after this change i can't detect races locally with the race detector.