slow down replay emits if outbox is getting filled
hopefully reduce occurrences of https://github.com/bluesky-social/jetstream/issues/27 / https://github.com/bluesky-social/jetstream/issues/31

when replaying without any filters, two rate-limiters are active: the db rate limiter will allow max-sub-rate*10 events to be added to the client's queue, and the emitting code will allow max-sub-rate to be emitted to the client.

is _seems likely_ that this should lead to a nearly-full outbox as a fairly normal scenario during replay. a nearly-full outbox puts clients at high risk of getting booted for being "slow" as soon as the cutover to live-tail happens.

this change is a hack. if emitting from replay it will just start slowwwwwing down by sleeping for 100ms at a time whenever trying to emit to the subscriber outbox with that outbox over 1/3 of its capacity.

the go race still seems happy with this change, my own out-of-order detecting script shows *maybe*, *marginally* better behaviour in terms of out-of-ordered emits. i don't have a good repro for the prod cutover client booting, but with luck this might help.